requirements.txt
click==8.1.3
PyYAML==6.0
# Editable install with no version control (sshmanager==0.1.0)
-e /Users/craigeaveson/Documents/Dev/sshmanager
